07-项目实战篇/股票数据监控系统/stock_monitor/pyproject.toml
[project]
name = "stock-monitor"
version = "0.1.0"
description = "股票数据监控系统 - 股票行情采集与分析"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]